ai_career_advisor_ai_powered/requirements.txt
streamlit
pandas
transformers
torch
sentence-transformers
python-dotenv
